Impelement add profile picture, manage role, navbar restriction and error feadback

+++ app/Http/Controllers/API/AsistenController.php
@@ -2,38 +2,46 @@
 
 namespace App\Http\Controllers\API;
 
+use ImageKit\ImageKit;
 use App\Models\Asisten;
+use App\Models\FotoAsisten;
 use Illuminate\Http\Request;
+use App\Adapter\ImageKitAdapter;
 use Spatie\Permission\Models\Role;
 use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Hash;
+use Intervention\Image\ImageManager;
+use Illuminate\Support\Facades\Validator;
+use Intervention\Image\Drivers\Gd\Driver;
+use Intervention\Image\Encoders\WebpEncoder;
 
 class AsistenController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
         try {
-            // Ambil data asisten dari database
-            $asisten = Asisten::leftJoin('roles', 'roles.id', '=', 'asistens.role_id')
-                ->select('nama', 'kode', 'roles.name as role', 'role_id', 'nomor_telepon', 'id_line', 'instagram', 'deskripsi')
-                ->get();
-
+            if(!auth()->guard('asisten')->user()) {
+                $asisten = Asisten::leftJoin('foto_asistens', 'foto_asistens.kode', '=', 'asistens.kode')
+                    ->select('nama', 'asistens.kode', 'foto_asistens.foto', 'nomor_telepon', 'id_line', 'instagram', 'deskripsi')
+                    ->orderBy('asistens.kode', 'asc')->get(); 
+                
+            }else{
+                $asisten = Asisten::leftJoin('roles', 'roles.id', '=', 'asistens.role_id')
+                ->leftJoin('foto_asistens', 'foto_asistens.kode', '=', 'asistens.kode')
+                ->select('nama', 'asistens.kode', 'foto_asistens.foto', 'roles.name as role', 'role_id', 'nomor_telepon', 'id_line', 'instagram', 'deskripsi')
+                ->orderBy('asistens.kode', 'asc')->get();
+            }
             return response()->json([
                 'success' => true,
                 'asisten' => $asisten,
                 'message' => 'Asisten retrieved successfully.',
             ], 200);
         } catch (\Exception $e) {
-            return response()->json([
-                'success' => false,
-                'message' => 'Failed to retrieve Asisten.',
-                'error' => $e->getMessage(),
-            ], 500);
+            return redirect()->back()->with('error', 'Failed to retrieve Asisten.');
         }
     }
 
@@ -90,6 +98,98 @@ class AsistenController extends Controller
         }
     }
 
+    public function updatePp(Request $request){
+        // //inii bedaaa lagi
+        if ($request->hasFile('upload')) {
+            try {
+                $validator = Validator::make($request->all(), [
+                    'upload' => 'image|mimes:jpeg,png,jpg|max:512',
+                ]);
+
+                if ($validator->fails()) {
+                    return back()->withErrors($validator->errors()); // Inertia handles errors automatically
+                }
+
+                $imageKit = new ImageKit(
+                    env('IMAGEKIT_PUBLIC_KEY'),
+                    env('IMAGEKIT_PRIVATE_KEY'),
+                    env('IMAGEKIT_ENDPOINT_URL')
+                );
+
+                $user = auth('asisten')->user();
+                $file = $request->file('upload');
+
+                // Find existing record
+                $foto = FotoAsisten::where('kode', $user->kode)->first();
+
+                // Initialize adapter
+                $adapter = new ImageKitAdapter($imageKit);
+
+                // If an old image exists, delete it from ImageKit
+                if ($foto && $foto->foto) {
+                    $adapter->delete($foto->fileId);
+                }
+
+            
+                // Upload to ImageKit
+                $result = $adapter->upload('ProfilePicDaskom/' . $user->kode . '.' . $file->getClientOriginalExtension(), $file)->result;
+
+                // Create or update FotoAsisten entry
+                if ($foto) {
+                    $foto->update([
+                        'foto' => env('IMAGEKIT_ENDPOINT_URL') . $result->filePath. '?updatedAt='. now()->timestamp * 1000,
+                        'fileId' => $result->fileId,
+                    ]);
+                } else {
+                    FotoAsisten::create([
+                        'kode' => $user->kode,
+                        'foto' => env('IMAGEKIT_ENDPOINT_URL') . $result-> filePath . '?updatedAt=' . now()->timestamp * 1000,
+                        'fileId' => $result->fileId,
+                        'created_at' => now(),
+                        'updated_at' => now(),
+                    ]);
+                }
+
+                return back()->with('success', 'Profile picture updated successfully.');
+            } catch (\Exception $e) {
+                return back()->with('error', 'Something went wrong: ' . $e->getMessage());
+            }
+        }
+
+        return back()->with('error', 'No file uploaded.');
+        
+    }
+
+
+    public function destroyPp(){
+
+        try {
+            $imageKit = new ImageKit(
+                env('IMAGEKIT_PUBLIC_KEY'),
+                env('IMAGEKIT_PRIVATE_KEY'),
+                env('IMAGEKIT_ENDPOINT_URL')
+            );
+
+            $user = auth('asisten')->user();
+
+            // Find existing record
+            $foto = FotoAsisten::where('kode', $user->kode)->first();
+
+            // If an old image exists, delete it from ImageKit
+            $adapter = new ImageKitAdapter($imageKit);
+
+            if ($foto && $foto->foto) {
+                $adapter->delete($foto->fileId);
+                FotoAsisten::where('kode', $user->kode)->delete();
+            }
+
+            return back()->with('success', 'Profile picture deleted successfully.');
+        } catch (\Exception $e) {
+            return back()->with('error', 'Failed to delete profile picture: ' . $e->getMessage());
+        }
+
+    }
+
 
 
     /**
@@ -133,22 +233,32 @@ class AsistenController extends Controller
     }
 
 
-
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy($id)
+    public function destroy(Request $request)
     {
         try {
-            $asisten = Asisten::findOrFail($id);
-            $role = Role::findOrFail($asisten->role_id);
-            $asisten->removeRole($role->name);
+        $data = request()->input('asistens', []);
+
+        if (empty($data)) {
+            return redirect(route('manage-role'))->with('error', 'No data provided.');
+        }
+
+        foreach ($data as $kode) {
+            $asisten = Asisten::where('kode', $kode)->first();
+
+            if (!$asisten) {
+                return redirect(route('manage-role'))->with('error', "Asisten with kode $kode not found.");
+            }
+
+            $role = Role::find($asisten->role_id);
+            if ($role) {
+                $asisten->removeRole($role->name);
+            }
+
             $asisten->delete();
-            return response()->json([
-                'success' => true,
-                'message' => 'Asisten deleted successfully.',
-            ], 200);
+        }
+
+        // Return success message
+        return redirect(route('manage-role'))->with('success', 'Selected Asistens deleted successfully.');
         } catch (\Illuminate\Database\Eloquent\ModelNotFoundException $e) {
             return response()->json([
                 'success' => false,
